Finished implementation and tests on (simple) trace data-structure
I have a few ideas to improve this data-structure, but I will do it later.
The idea is to group an array of pointers in each node like this:

\#define TNODE_SIZE 1ULL << 8

typedef struct tnode_t
{
  instr_t *instr[TNODE_SIZE];
  struct tnode_t *next;
} tnode_t;

It should reduce the number of calls to malloc() and the size of the whole
structure in memory. We first fill the array with NULL (search for the last
non empty element is performed by dichotomic search).